moved DefaultClient to model
